Change bracket strip
This is more reliable, as we first test to see if there are any brackets in front BEFORE stripping the brackets. This way, we don't shave off text that Slack does not recognize as links. People should be sending in links, but this is an optional check in case something goes wrong.