Fix timeout issue
The websocket connection drops because of inactivity. When it does, the slackclient library does not automatically attempt to reconnect without explicitly being told to do so. When we try to send a message while the socket is closed, we get an error and the script crashes.

Pages related to this issue:
https://github.com/slackapi/python-slackclient/issues/118
https://slack.dev/python-slackclient/real_time_messaging.html#rtm-start-vs-rtm-connect

Possible auto-ping solution required:
https://github.com/TheKevJames/jarvis/blob/master/bot/jarvis/core/async.py#L39-L49

-----------------------------

Turns out, the websocket-client library has a fault with 0.55.0. Downgrading to 0.54.0 *should* fix the issue, but I already have 0.54.0 on my local machine, so that's not the problem.

Issues:
https://github.com/slackapi/python-slackclient/issues/385
https://github.com/slackapi/python-slackclient/issues/127
https://github.com/slackapi/python-slackclient/pull/387